Bullets using Entity Class
Thought i'd upload this version for now incase people need it: I've not
got any animations for the Bullet and ASSUMING we're going to have a
sprite sheet for it, I've borrowed the Mummy's Animations for now. So
although the code is right (for a 16x16 Bullet!) we still need some
bullet sprites and the asset path changing.