Artificial Unintelligence Added
A look into some proper node pathing (found out about A* today) might be
good for the future. For now, the mummy will at least chase you if you
stay within a couple of blocks. If we set it to a larger range than that
he gets stuck in corners and such.